python/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-google-genai==2.0.1
google-generativeai==0.8.3
python-dotenv==1.0.1
langgraph==0.2.45
langchain-core==0.3.15
langchain-postgres==0.0.12
langchain-text-splitters==0.3.2
psycopg[binary]==3.1.18